broke out mongodb document models to a separate file
